+++ packages/satteri/package.json
@@ -50,6 +50,7 @@
     "estree-util-to-js": "^2.0.0",
     "estree-util-value-to-estree": "^3.5.0",
     "fumadocs-core": "workspace:*",
+    "fumadocs-typescript": "workspace:*",
     "github-slugger": "^2.0.0",
     "hast-util-to-estree": "^3.1.3",
     "image-size": "^2.0.2",
