+++ packages/typescript/src/index.ts
@@ -1,6 +1,6 @@
 export * from '@/lib/base';
 export type { MarkdownRenderer } from './markdown';
 export * from '@/lib/remark-auto-type-table';
-export { remarkAutoTypeTableSatteri } from '@/lib/remark-auto-type-table';
+export * from '@/lib/auto-type-table-generate';
 export * from './cache/fs-cache';
 export * from './cache';
