+++ packages/typescript/src/lib/remark-auto-type-table.ts
@@ -1,151 +1,15 @@
 import type { Root } from 'mdast';
-import type { Nodes } from 'hast';
 import type { Transformer } from 'unified';
-import type { Expression, ExpressionStatement, ObjectExpression } from 'estree';
-import { createGenerator, type DocEntry, type Generator } from '@/lib/base';
-import { type MarkdownRenderer, markdownRenderer, type ShikiOptions } from '@/markdown';
-import { valueToEstree } from 'estree-util-value-to-estree';
 import { visit } from 'unist-util-visit';
-import { type BaseTypeTableProps, type GenerateTypeTableOptions } from '@/lib/type-table';
-import { toEstree } from 'hast-util-to-estree';
-import { type ParameterTag, parseTags } from '@/lib/parse-tags';
-import type { MdxJsxAttribute, MdxJsxExpressionAttribute, MdxJsxFlowElement } from 'mdast-util-mdx';
+import type { MdxJsxAttribute, MdxJsxExpressionAttribute } from 'mdast-util-mdx';
 import type { VFile } from 'vfile';
-import { defineMdastPlugin } from 'satteri';
-import type { MdastNode } from 'satteri';
-import '@fumadocs/satteri/data-map';
-import { fileURLToPath } from 'node:url';
-import path from 'node:path';
+import {
+  createAutoTypeTableGenerator,
+  type RemarkAutoTypeTableOptions,
+  type TypeTableProps,
+} from '@/lib/auto-type-table-generate';
 
-function objectBuilder() {
-  const out: ObjectExpression = {
-    type: 'ObjectExpression',
-    properties: [],
-  };
-
-  return {
-    addExpressionNode(key: string, expression: Expression) {
-      out.properties.push({
-        type: 'Property',
-        method: false,
-        shorthand: false,
-        computed: false,
-        key: {
-          type: 'Literal',
-          value: key,
-        },
-        kind: 'init',
-        value: expression,
-      });
-    },
-    addJsxProperty(key: string, hast: Nodes) {
-      const estree = toEstree(hast, {
-        elementAttributeNameCase: 'react',
-      }).body[0] as ExpressionStatement;
-
-      this.addExpressionNode(key, estree.expression);
-    },
-    build() {
-      return out;
-    },
-  };
-}
-
-async function buildTypeProp(
-  entries: DocEntry[],
-  renderer: MarkdownRenderer,
-): Promise<ObjectExpression> {
-  async function onItem(entry: DocEntry) {
-    const node = objectBuilder();
-    const tags = parseTags(entry.tags);
-    node.addJsxProperty('type', await renderer.renderTypeToHast(entry.simplifiedType));
-    node.addJsxProperty('typeDescription', await renderer.renderTypeToHast(entry.type));
-    node.addExpressionNode('required', valueToEstree(entry.required));
-
-    if (entry.typeHref)
-      node.addExpressionNode('typeDescriptionLink', valueToEstree(entry.typeHref));
-
-    if (tags.default) node.addJsxProperty('default', await renderer.renderTypeToHast(tags.default));
-
-    if (tags.returns)
-      node.addJsxProperty('returns', await renderer.renderMarkdownToHast(tags.returns));
-
-    if (tags.params) {
-      node.addExpressionNode('parameters', {
-        type: 'ArrayExpression',
-        elements: await Promise.all(tags.params.map(onParam)),
-      });
-    }
-
-    if (entry.description) {
-      node.addJsxProperty('description', await renderer.renderMarkdownToHast(entry.description));
-    }
-
-    if (entry.deprecated) {
-      node.addExpressionNode('deprecated', valueToEstree(true));
-    }
-
-    return node.build();
-  }
-
-  async function onParam(param: ParameterTag) {
-    const node = objectBuilder();
-    node.addExpressionNode('name', valueToEstree(param.name));
-    if (param.description)
-      node.addJsxProperty('description', await renderer.renderMarkdownToHast(param.description));
-
-    return node.build();
-  }
-
-  const prop = objectBuilder();
-  const output = await Promise.all(
-    entries.map(async (entry) => ({
-      name: entry.name,
-      node: await onItem(entry),
-    })),
-  );
-
-  for (const node of output) {
-    prop.addExpressionNode(node.name, node.node);
-  }
-
-  return prop.build();
-}
-
-export interface RemarkAutoTypeTableOptions {
-  /**
-   * @defaultValue 'auto-type-table'
-   */
-  name?: string;
-
-  /**
-   * @defaultValue 'TypeTable'
-   */
-  outputName?: string;
-
-  /**
-   * config for Shiki when using default `renderMarkdown` & `renderType`.
-   */
-  shiki?: ShikiOptions;
-  renderMarkdown?: MarkdownRenderer['renderMarkdownToHast'];
-  renderType?: MarkdownRenderer['renderTypeToHast'];
-
-  /**
-   * Customize type table generation
-   */
-  options?: GenerateTypeTableOptions;
-
-  /**
-   * generate the stringified form of props (useful for `remark-stringify` etc).
-   */
-  remarkStringify?: boolean;
-
-  generator?: Generator;
-}
-
-export interface TypeTableProps extends BaseTypeTableProps {
-  cwd?: true;
-}
+export type { RemarkAutoTypeTableOptions, TypeTableProps } from '@/lib/auto-type-table-generate';
 
 /**
  * Compile `auto-type-table` into Fumadocs UI compatible TypeTable
@@ -155,88 +19,13 @@ export interface TypeTableProps extends BaseTypeTableProps {
 export function remarkAutoTypeTable(
   config: RemarkAutoTypeTableOptions = {},
 ): Transformer<Root, Root> {
-  const {
-    name = 'auto-type-table',
-    outputName = 'TypeTable',
-    options: generateOptions = {},
-    remarkStringify = true,
-    generator = createGenerator(),
-    renderMarkdown,
-    renderType,
-    shiki,
-  } = config;
-  let renderer: MarkdownRenderer;
-
-  if (renderMarkdown && renderType) {
-    renderer = { renderMarkdownToHast: renderMarkdown, renderTypeToHast: renderType };
-  } else {
-    renderer = markdownRenderer(shiki);
-    if (renderMarkdown) renderer.renderMarkdownToHast = renderMarkdown;
-    if (renderType) renderer.renderTypeToHast = renderType;
-  }
-
-  async function generate(
-    file: VFile,
-    props: TypeTableProps,
-    attributes: (MdxJsxAttribute | MdxJsxExpressionAttribute)[],
-  ) {
-    let basePath = props.cwd ? file.cwd : generateOptions.basePath;
-    if (file.dirname) {
-      basePath ??= file.dirname;
-    }
-
-    const output = await generator.generateTypeTable(props, {
-      ...generateOptions,
-      basePath,
-    });
-    const rendered: MdxJsxFlowElement[] = [];
-
-    for (const doc of output) {
-      rendered.push({
-        type: 'mdxJsxFlowElement',
-        name: outputName,
-        attributes: [
-          {
-            type: 'mdxJsxAttribute',
-            name: 'id',
-            value: `type-table-${doc.id}`,
-          },
-          {
-            type: 'mdxJsxAttribute',
-            name: 'type',
-            value: {
-              type: 'mdxJsxAttributeValueExpression',
-              value: remarkStringify ? JSON.stringify(doc, null, 2) : '',
-              data: {
-                estree: {
-                  type: 'Program',
-                  sourceType: 'module',
-                  body: [
-                    {
-                      type: 'ExpressionStatement',
-                      expression: await buildTypeProp(doc.entries, renderer),
-                    },
-                  ],
-                },
-              },
-            },
-          },
-          ...attributes,
-        ],
-        children: [],
-      });
-    }
-
-    return rendered;
-  }
+  const { name, generateTables } = createAutoTypeTableGenerator(config);
 
   return async (tree, file) => {
     const queue: Promise<void>[] = [];
 
     visit(tree, 'mdxJsxFlowElement', (node) => {
       if (node.name !== name) return;
-      const props: TypeTableProps = {};
-      const attributes: (MdxJsxAttribute | MdxJsxExpressionAttribute)[] = [];
 
       const onError = (message: string, cause?: Error) => {
         const location = node.position
@@ -246,6 +35,9 @@ export function remarkAutoTypeTable(
           cause,
         });
       };
+
+      const props: TypeTableProps = {};
+      const attributes: (MdxJsxAttribute | MdxJsxExpressionAttribute)[] = [];
 
       for (const attr of node.attributes) {
         if (attr.type !== 'mdxJsxAttribute') {
@@ -274,7 +66,7 @@ export function remarkAutoTypeTable(
       }
 
       queue.push(
-        generate(file, props, attributes)
+        generateTables(file.path, props, attributes, file.cwd)
           .then((children) => {
             Object.assign(node, {
               type: 'root',
@@ -290,134 +82,4 @@ export function remarkAutoTypeTable(
 
     await Promise.all(queue);
   };
-}
-
-export function remarkAutoTypeTableSatteri(config: RemarkAutoTypeTableOptions = {}) {
-  const {
-    name = 'auto-type-table',
-    renderMarkdown,
-    renderType,
-    shiki,
-    generator = createGenerator(),
-  } = config;
-
-  let renderer: MarkdownRenderer;
-  if (renderMarkdown && renderType) {
-    renderer = { renderMarkdownToHast: renderMarkdown, renderTypeToHast: renderType };
-  } else {
-    renderer = markdownRenderer(shiki);
-    if (renderMarkdown) renderer.renderMarkdownToHast = renderMarkdown;
-    if (renderType) renderer.renderTypeToHast = renderType;
-  }
-
-  async function generateTables(
-    filePath: string | undefined,
-    props: TypeTableProps,
-    attributes: (MdxJsxAttribute | MdxJsxExpressionAttribute)[],
-    projectCwd?: string,
-  ) {
-    const {
-      outputName = 'TypeTable',
-      options: generateOptions = {},
-      remarkStringify = true,
-    } = config;
-
-    let basePath = props.cwd ? projectCwd : generateOptions.basePath;
-    if (filePath) basePath ??= path.dirname(filePath);
-
-    const output = await generator.generateTypeTable(props, {
-      ...generateOptions,
-      basePath,
-    });
-    const rendered: MdxJsxFlowElement[] = [];
-
-    for (const doc of output) {
-      rendered.push({
-        type: 'mdxJsxFlowElement',
-        name: outputName,
-        attributes: [
-          {
-            type: 'mdxJsxAttribute',
-            name: 'id',
-            value: `type-table-${doc.id}`,
-          },
-          {
-            type: 'mdxJsxAttribute',
-            name: 'type',
-            value: {
-              type: 'mdxJsxAttributeValueExpression',
-              value: remarkStringify ? JSON.stringify(doc, null, 2) : '',
-              data: {
-                estree: {
-                  type: 'Program',
-                  sourceType: 'module',
-                  body: [
-                    {
-                      type: 'ExpressionStatement',
-                      expression: await buildTypeProp(doc.entries, renderer),
-                    },
-                  ],
-                },
-              },
-            },
-          },
-          ...attributes,
-        ],
-        children: [],
-      });
-    }
-
-    return rendered;
-  }
-
-  return defineMdastPlugin({
-    name: 'remark-auto-type-table',
-    async mdxJsxFlowElement(node, ctx) {
-      if (node.name !== name) return;
-
-      const props: TypeTableProps = {};
-      const attributes: (MdxJsxAttribute | MdxJsxExpressionAttribute)[] = [];
-      const filePath = ctx.fileURL ? fileURLToPath(ctx.fileURL) : undefined;
-
-      for (const attr of node.attributes) {
-        if (attr.type !== 'mdxJsxAttribute') {
-          attributes.push(attr);
-          continue;
-        }
-
-        switch (attr.name) {
-          case 'cwd':
-            props.cwd = true;
-            break;
-          case 'path':
-          case 'name':
-          case 'type':
-            if (typeof attr.value === 'string') props[attr.name] = attr.value;
-            break;
-          default:
-            attributes.push(attr);
-        }
-      }
-
-      const parent = ctx.parent(node);
-      const index = ctx.indexOf(node);
-      if (!parent || index === undefined) return;
-
-      const children = await generateTables(
-        filePath,
-        props,
-        attributes,
-        ctx.data._cwd,
-      );
-      if (children.length === 0) {
-        ctx.removeNode(node);
-      } else if (children.length === 1) {
-        ctx.replaceNode(node, children[0]!);
-      } else {
-        const index = ctx.indexOf(node)!;
-        ctx.replaceNode(node, children[0]!);
-        ctx.insertChildAt(parent, index + 1, children.slice(1));
-      }
-    },
-  });
 }
